Added wonders setup at the beginning of the game

+++ ProjectModernCpp/ProjectModernCpp/Building.cpp
@@ -1,10 +1,10 @@
 module Building;
 
 Building::Building(const std::vector<Effect>& effects,
-	const Cost& cost, std::string_view name, std::uint8_t id,
-	std::uint8_t victoryPoints,
-	std::uint8_t shields,
-	std::uint8_t coins,
+	const Cost& cost, std::string_view name, std::uint16_t id,
+	std::uint16_t victoryPoints,
+	std::uint16_t shields,
+	std::uint16_t coins,
 	Age age,
 	Color color,
 	const std::vector<ResourceType>& resources,
